requirements.txt
# Core
chromadb>=0.5.0
sentence-transformers>=2.7.0
ollama>=0.2.0

# Agents & Skills
duckduckgo-search>=6.2.0
trafilatura>=1.9.0
arxiv>=2.1.0
youtube-transcript-api>=0.6.0
pytube>=15.0.0
pymupdf>=1.24.0

# API & Scheduler
fastapi>=0.111.0
uvicorn>=0.30.0
apscheduler>=3.10.0

# Utils
pyyaml>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.7.0
rich>=13.7.0
loguru>=0.7.0
httpx>=0.27.0

# Optional
anthropic>=0.30.0
